Rename test dir to allude sig-apps ownership
