+++ hack/.golint_failures
@@ -801,6 +801,7 @@ staging/src/k8s.io/sample-apiserver/pkg/registry/wardle/fischer
 staging/src/k8s.io/sample-apiserver/pkg/registry/wardle/flunder
 test/e2e
 test/e2e/apimachinery
+test/e2e/apps
 test/e2e/autoscaling
 test/e2e/chaosmonkey
 test/e2e/common
@@ -817,7 +818,6 @@ test/e2e/scheduling
 test/e2e/storage
 test/e2e/upgrades
 test/e2e/upgrades/apps
-test/e2e/workload
 test/e2e_federation
 test/e2e_federation/framework
 test/e2e_federation/upgrades
