+++ test/e2e/apimachinery/BUILD
@@ -25,9 +25,9 @@ go_library(
         "//pkg/controller:go_default_library",
         "//pkg/printers:go_default_library",
         "//pkg/util/version:go_default_library",
+        "//test/e2e/apps:go_default_library",
         "//test/e2e/framework:go_default_library",
         "//test/e2e/metrics:go_default_library",
-        "//test/e2e/workload:go_default_library",
         "//test/utils:go_default_library",
         "//vendor/github.com/onsi/ginkgo:go_default_library",
         "//vendor/github.com/onsi/gomega:go_default_library",
