+++ test/e2e/apps/types.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package workload
+package apps
 
 import (
 	"k8s.io/apimachinery/pkg/runtime/schema"
